moving Material.alpha into LLG integrator and Material.T into StochIntegrator

+++ scripts/bash/build.sh
@@ -5,8 +5,10 @@
 currdir=$PWD
 if [ -d $1/bin ]
 then 
-    [ "$2" == "true" ] && echo "removing files in $1/bin/*"
-    rm $1/bin/*
+    if [ -z "(ls -A $1/bin)" ]; then
+        [ "$2" == "true" ] && echo "removing files in $1/bin/*"
+        rm $1/bin/*
+    fi
 fi
 if [ ! -d $1/build ]
 then
